+++ app/CHANGELOG.md
@@ -3,6 +3,7 @@
 * **Version 0.1 22-11-12** Added the ability to add posts to website
 
   * Added the ability to add posts to website
+  
 * **Version 0.11 22-11-13**
 
   * Added library for definitions
@@ -11,11 +12,13 @@
   * Added the ability to check changelog and cleanup files (New library `UtilsLib.h`)
   * Added new function in `FileLib.h` that calculates the size of string in buffer
     (I think it already exists in common libraries, but I forgot to check).
+    
 * **Version 0.12 22-11-13 (A.K.A Losing sanity update)**
 
   * Added the ability to remove posts!
     It removes it from the log file and from the `/data` folder
     Added handy `SEPARATOR` and `SEPARATORIND` defines
+    
 * **Version 0.22 22-11-17 (A.K.A Crying in the corner update)**
 
   * Completely reworked code **IGNORE PAST UPDATES**
@@ -26,6 +29,7 @@
   * Added a define to remove the ending `\n` symbol
   * Added the creation of log files for each category
   * And a whole lot more...
+  
 * **Version 0.3 22-12-02**
 
   * Added message creation and storing in files in their html formated form
@@ -34,24 +38,34 @@
   * **v0.1** of building page file
     * Loads all of standard templates and category information into one html file sucessfully
   * Many other additions and improvements overall...
+  
 * **Version 0.5 Some Date**
 
   * So much that we forgot to list and forgot what we did, basically made it compile itself and other
+  
 * **Version 0.6 22-12-08**
 
   * Added the option to delete categories and messages
   * Also deletes all log files and page 60% of the time
+  
 * **Version 0.7 22-12-11 (a.k.a. first version which can be used properly)**
 
   * Nav bar gets updated with newer categories
   * Minor changes and bugfixes to smoothen the process a little
+  
 * **Version 0.7.1 22-12-3rd week (a.k.a. "polishing")**
 
   * Input validation
   * Less hardcoding
   * Ability to quit at some stages
   * Less crashing (probably not)
+  
 * **Version 0.7.2 22-12-26**
 
   * Fixed issue when not being able to delete decription log just after creating it
   * Fixed oopsie v2
+
+* **Version 0.7.5 22-12-31**
+
+  * Now a duplicate category or message cannot be created
+  * Also you can quit from creating a message title